Data encryption hook
